chore(products): move products into its own folder
