+++ src/pages/BrowserPage/Products/ProductsGrid.jsx
