+++ src/pages/BrowserPage/Products/ViewModeToggle.jsx
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 import { Button } from '@ynput/ayon-react-components'
